More refactoring and added support for solid meshing of Pipe wrt to FEM mesh creation.
